requirements_test.txt
-r ./requirements.txt
pylint>=2.3.1
pylint-django>=2.0.11
coverage>=4.5.4
django-coverage>=1.2.4
django-coverage-plugin>=1.6.0
